tweaks

+++ app/src/main/java/nars/NAgentX.java
@@ -89,9 +89,12 @@ abstract public class NAgentX extends NAgent {
         return runRT(init, fps, 1, -1);
     }
 
-    public static NAR runRT(Function<NAR, NAgent> init, float fps, int durFrames, int endTime) {
+    public static NAR runRT(Function<NAR, NAgent> init, float fps, float durFrames, int endTime) {
 
-        Time clock = new RealTime.DSHalf(true).durSeconds(durFrames / fps);
+        Time clock = new RealTime.
+                DSHalf(true)
+                //CS(true)
+                .durSeconds(durFrames / fps);
         Default nar =
                 //new TaskNAR(32 * 1024, new MultiThreadExecutioner(4, 4 * 1024), clock);
                 NARBuilder.newMultiThreadNAR(-1, clock, true);
